chore(core): nx reset command, updates to cli help output and docs
