+++ docs/node/cli/reset.md
@@ -1,11 +1,11 @@
-# clear-cache
+# reset
 
-Clears all the cached Nx artifacts and metadata about the workspace.
+Clears all the cached Nx artifacts and metadata about the workspace and shuts down the Nx Daemon.
 
 ## Usage
 
 ```bash
-nx clear-cache
+nx reset
 ```
 
 [Install `nx` globally]({{framework}}/getting-started/nx-setup#install-nx) to invoke the command directly using `nx`, or use `npx nx`, `yarn nx`, or `pnpx nx`.
