+++ packages/workspace/src/core/project-graph/daemon/client/generate-help-output.ts
@@ -2,7 +2,13 @@ import { spawnSync } from 'child_process';
 import { getDaemonProcessId } from '../cache';
 import { DAEMON_OUTPUT_LOG_FILE } from '../tmp-dir';
 
-export function generateDaemonHelpOutput(): string {
+export function generateDaemonHelpOutput(
+  isGenerateDocsProcess = false
+): string {
+  if (isGenerateDocsProcess) {
+    return `The Nx Daemon is a local server which runs in the background in order to intelligently cache information about the workspace's project graph.`;
+  }
+
   /**
    * A workaround for cases such as yargs output where we need to synchronously
    * get the value of this async operation.
