+++ packages/workspace/src/core/project-graph/daemon/server/server.ts
@@ -274,7 +274,6 @@ export async function stopServer(): Promise<void> {
       }
 
       killSocketOrPath();
-      logger.info('NX Daemon Server - Stopped');
       return resolve();
     });
   });
